Revert "safer tempfile choice" (commit 411ae24bbedfb0bc4ba651c7d40507d4118d33da.)
This seems to be the root cause of issue #19: on many devices, `mktemp -d`
returns a directory in `/data/local/tmp`.

This causes the transfer to freeze for the `userdata` partition: when
tetherback remounts `/data` as read-only, suddenly the `md5in` and `md5out`
files are on a read-only partition, and so I/O is blocked for `md5sum`.

Derp derp derp… I caused a really confusing bug by trying to make something
safer and more reliable :-(